\name{cronbach}
\alias{cronbach}
\title{Cronbach alpha}
\description{
to do
}
\usage{
cronbach(x)
}
\arguments{
    \item{x}{A data frame in which the rows represents the subjects and the columns the items.}
}
\details{
Missing data are treated by using listwise deletion.
}
\author{
Davide Massidda \email{davide.massidda@gmail.com}
}
\seealso{
\code{\link[raschTools]{reliability}}
}
\examples{
# to do
}
