\name{reliability}
\alias{reliability}
\alias{print.reliability}
\title{Internal Consistence}
\description{
to do
}
\usage{
reliability(x, cor.method = c("pearson","biserial","polyserial"), fn = sum)
}
\arguments{
    \item{x}{to do}
    \item{cor.method}{to do}
    \item{fn}{to do}
}
\details{
to do
}
\author{
Davide Massidda \email{davide.massidda@gmail.com}
}
\seealso{
\code{\link[raschTools]{cronbach}}, \code{\link[raschTools]{rmfit}},
}
\examples{
# to do
}
