\name{rmfit}
\alias{rmfit}
\alias{print.rmfit}
\title{Fit indices for Rasch models}
\description{
to do
}
\usage{
rmfit(model, person.par = NULL)
}
\arguments{
    \item{model}{to do}
    \item{person.par}{to do}
}
\details{
to do
}
\author{
Davide Massidda \email{davide.massidda@gmail.com}
}
\seealso{
\code{\link[eRm]{PCM}}
}
\examples{
# to do
}
