\name{simSLM}
\alias{simPCM}
\alias{RaschData-class}
\title{Simulate Rasch Models}
\description{
Person-item map for the Partial Credit Model estimated with the function \code{PCM} of the \code{eRm} package.
}
\usage{
simSLM(numSubj = 10, numItem = 5, seed = NULL)
simPCM(numSubj = 10, numItem = 5, maxScore = 3, seed = NULL)
\arguments{
    \item{numSubj}{to do}
    \item{numItem}{to do}
    \item{maxScore}{to do}
    \item{seed}{to do}
}
\details{
to do
}
\author{
Davide Massidda \email{davide.massidda@gmail.com}
}
\seealso{
\code{\link[eRm]{RM}}, \code{\link[eRm]{PCM}}
}
\examples{\dontrun{
# to do
}}
